refactor(analysis): extract API contract validation; internal/analysis is now a directory
Last of the four disjoint subsystems. The api_contract_* files move to
internal/analysis/apicontract with the prefix dropped, taking contract_test.go
and what remains of the package-wide branch_coverage_test.go (describeType,
toStringMap, recordObservation — all api-contract code).

With that, internal/analysis holds no source of its own, so doc.go goes too:
its four bullet points are now package docs on the packages they describe. The
directory drops out of .folder-size-baseline.json entirely (28 -> 0), and the
four packages are 6/7/6/5 files — all under the limit with headroom.

binary_integration_test.go was misfiled, not a schema test: it constructs a
capture.Capture, pushes network bodies and WebSocket events through it, and
asserts binary_format/format_confidence — it never touches an analysis symbol.
It also sat behind `//go:build integration`, and the stated reason ("Run with:
go test -tags=integration") is not a justification: it needs no external
resource, no network and no fixture. Moved to internal/capture as
binary_format_integration_test.go (package capture_test) with the tag removed,
so five real assertions about ingest-time format detection now actually run.
Verified passing, including under -race.

Consumers updated: tools_analyze_api_validation.go, tools_core.go,
tools_core_constructor.go.

Coverage of the new package is 92.8%. One function reads 0.0%,
detectTopLevelTypeChange — pre-existing, not caused by the move: it measures
0.0% on origin/UNSTABLE too, because compareShapes only reaches it when a
learned shape is not a map and extractShape always produces one. Reported as a
finding rather than papered over with a test.
