+++ .folder-size-baseline.json
@@ -7,7 +7,6 @@
     "cmd/browser-agent/internal/toolconfigure": 14,
     "cmd/browser-agent/internal/toolgenerate": 13,
     "cmd/browser-agent/internal/toolinteract": 42,
-    "internal/analysis": 28,
     "internal/capture": 67,
     "internal/checkpoint": 11,
     "internal/mcp": 12,
