+++ cmd/browser-agent/tools_core.go
@@ -15,7 +15,7 @@ import (
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/testgenhandler"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolconfigure"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolinteract"
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/analysis"
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/analysis/apicontract"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/analysis/thirdparty"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/audit"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/capture"
@@ -79,7 +79,7 @@ type ToolHandler struct {
 
 	// API contract validation state (incremental over captured network bodies).
 	apiContractMu        sync.Mutex
-	apiContractValidator *analysis.APIContractValidator
+	apiContractValidator *apicontract.APIContractValidator
 	apiContractOffset    int
 
 	// Upload security config (folder-scoped permissions + denylist)
