+++ docs/features/feature/api-schema/index.md
@@ -7,12 +7,12 @@ owners: []
 last_reviewed: 2026-07-26
 code_paths:
   - cmd/browser-agent/openapi.go
-  - internal/analysis/api_contract.go
-  - internal/analysis/api_contract_analysis.go
-  - internal/analysis/api_contract_endpoint.go
-  - internal/analysis/api_contract_learning.go
-  - internal/analysis/api_contract_validation.go
-  - internal/analysis/api_contract_violations.go
+  - internal/analysis/apicontract/contract.go
+  - internal/analysis/apicontract/report.go
+  - internal/analysis/apicontract/endpoint.go
+  - internal/analysis/apicontract/learning.go
+  - internal/analysis/apicontract/validation.go
+  - internal/analysis/apicontract/violations.go
   - internal/analysis/apischema/schema.go
   - internal/analysis/apischema/builder.go
   - internal/analysis/apischema/openapi.go
@@ -22,8 +22,8 @@ code_paths:
   - internal/analysis/apischema/observe_ws.go
   - internal/schema/observe.go
 test_paths:
-  - internal/analysis/api_contract_test.go
-  - internal/analysis/branch_coverage_test.go
+  - internal/analysis/apicontract/contract_test.go
+  - internal/analysis/apicontract/branch_coverage_test.go
   - internal/analysis/apischema/builder_test.go
   - internal/analysis/apischema/infer_test.go
   - internal/analysis/apischema/observe_http_test.go
